backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
pandas==2.2.2
numpy==2.1.1
python-dotenv==1.0.1
